Fix server commands from IRC
The dispatchCommand() method is not thread safe, and therefore should be
wrapped in a runnable and dispatched from a synchronized task. As well,
before, executing a command did not use getCommandSender(), instead
parsing the IRC nickname into a CommandSender, which causes problems if
your IRC nick is not the same as your ingame nick.

I also removed Eclipse specific files and added them to the gitignore,
as they aren't required for compilation and only clutter the required
code